soc/integration/export: Add CSR aliases for devices with suffix zero
The LiteX BIOS checks for the presence of the various CSR_*_BASE
definitions to decide to include device support or not.  E.g. for I2C,
it checks for CSR_I2C_BASE.

However, linux-on-litex-vexriscv registers I2C masters using "i2c0"
instead of "i2c", in anticipation for future support for multiple I2C
busses.  As a consequence, CSR_I2C0_BASE is defined instead of
CSR_I2C_BASE, and LiteX BIOS is built without support for I2C.

Fix this by creating aliases for devices with suffix zero in the
generated CSR register definitions, so the LiteX BIOS will pick up the
presence of the corresponding device, and include support for it.

For now this is limited to the first instance of each type.

Signed-off-by: Geert Uytterhoeven <geert@linux-m68k.org>